+++ app/scenes/Activity/index.js
@@ -463,7 +463,9 @@ class Activity extends React.Component {
             onChange={(answer, goToNext=false, timeElapsed=0) => {
               responses[currentScreen] = answer;
 
-              sendData('set_response', { [activity.items[currentScreen].id]: answer }, currentApplet.id);
+              if (this.currentItem.inputType != 'drawing' && this.currentItem.inputType != 'trail') {
+                sendData('set_response', { [activity.items[currentScreen].id]: answer }, currentApplet.id);
+              }
 
               this.setState({ responses })
               this.handleChange(answer, goToNext, timeElapsed);